# backend/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ML Evaluation Platform API",
    "description": "API for object detection and segmentation model evaluation platform",
    "version": "1.0.0",
    "contact": {
      "name": "ML Evaluation Platform"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.ml-eval.cloud",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "summary": "List available models",
        "description": "Retrieve list of available models for inference and training",
        "operationId": "listModels",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "detection",
                "segmentation"
              ]
            }
          },
          {
            "name": "framework",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "YOLO11",
                "YOLO12",
                "RT-DETR",
                "SAM2"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_id}": {
      "get": {
        "summary": "Get model details",
        "description": "Retrieve detailed information about a specific model",
        "operationId": "getModel",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/inference/single": {
      "post": {
        "summary": "Run single image inference",
        "description": "Run model inference on a single image",
        "operationId": "runSingleInference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SingleInferenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inference completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/inference/batch": {
      "post": {
        "summary": "Run batch inference",
        "description": "Run model inference on multiple images",
        "operationId": "runBatchInference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchInferenceRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Batch inference job created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceJob"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/inference/jobs/{job_id}": {
      "get": {
        "summary": "Get inference job status",
        "description": "Monitor batch inference job progress",
        "operationId": "getInferenceJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceJob"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Annotation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_id": {
            "type": "string",
            "format": "uuid"
          },
          "bounding_boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoundingBox"
            }
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Segment"
            }
          },
          "class_labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence_scores": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "creation_method": {
            "type": "string",
            "enum": [
              "user",
              "model"
            ]
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_tag": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "image_id",
          "class_labels",
          "creation_method",
          "created_at"
        ]
      },
      "BoundingBox": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          },
          "width": {
            "type": "number",
            "format": "float"
          },
          "height": {
            "type": "number",
            "format": "float"
          },
          "class_id": {
            "type": "integer"
          },
          "confidence": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ]
      },
      "Segment": {
        "type": "object",
        "properties": {
          "polygon": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "mask": {
            "type": "string",
            "description": "Base64 encoded mask"
          },
          "class_id": {
            "type": "integer"
          },
          "confidence": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "detection",
              "segmentation"
            ]
          },
          "variant": {
            "type": "string",
            "enum": [
              "nano",
              "small",
              "medium",
              "large",
              "xl"
            ]
          },
          "version": {
            "type": "string"
          },
          "framework": {
            "type": "string",
            "enum": [
              "YOLO11",
              "YOLO12",
              "RT-DETR",
              "SAM2"
            ]
          },
          "model_path": {
            "type": "string"
          },
          "training_status": {
            "type": "string",
            "enum": [
              "pre-trained",
              "training",
              "trained",
              "failed"
            ]
          },
          "performance_metrics": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "variant",
          "version",
          "framework",
          "model_path",
          "training_status",
          "created_at",
          "updated_at"
        ]
      },
      "ModelListResponse": {
        "type": "object",
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          },
          "total_count": {
            "type": "integer"
          }
        },
        "required": [
          "models",
          "total_count"
        ]
      },
      "SingleInferenceRequest": {
        "type": "object",
        "properties": {
          "image_id": {
            "type": "string",
            "format": "uuid"
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "confidence_threshold": {
            "type": "number",
            "format": "float",
            "default": 0.5
          }
        },
        "required": [
          "image_id",
          "model_id"
        ]
      },
      "BatchInferenceRequest": {
        "type": "object",
        "properties": {
          "image_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "confidence_threshold": {
            "type": "number",
            "format": "float",
            "default": 0.5
          }
        },
        "required": [
          "image_ids",
          "model_id"
        ]
      },
      "InferenceResult": {
        "type": "object",
        "properties": {
          "image_id": {
            "type": "string",
            "format": "uuid"
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "predictions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            }
          },
          "execution_time": {
            "type": "number",
            "format": "float"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "image_id",
          "model_id",
          "predictions",
          "execution_time",
          "timestamp"
        ]
      },
      "InferenceJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "target_images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ]
          },
          "progress_percentage": {
            "type": "number",
            "format": "float"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InferenceResult"
            }
          },
          "execution_logs": {
            "type": "string"
          },
          "start_time": {
            "type": "string",
            "format": "date-time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "model_id",
          "target_images",
          "status",
          "progress_percentage",
          "created_at"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error",
          "message",
          "timestamp"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}